requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
matplotlib==3.4.2
seaborn==0.11.2
joblib==1.0.1
google-generativeai==0.1.0
flask==2.0.1
gunicorn==20.1.0
-e .
